requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
cffi==1.14.3
cryptography==3.2
cycler==0.10.0
Django==3.1.2
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
wrapt==1.12.1
